fix: ghost release
